+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+owncloud-news (6.0.7)
+* **Enhancement**: Update articles if the pubdate is newer than the current one, #877
+
 owncloud-news (6.0.6)
 * **Enhancement**: Align settings text properly, #829
 * **Enhancement**: Update picoFeed to include additional feed rules